Module wakuchin::convert

Wakuchin conversion functions

Functions

chars_to_wakuchin

Convert from internally used wakuchin chars to actual wakuchin chars. This is useful when you want to display the wakuchin chars to the user.

wakuchin_to_chars

Convert from actual wakuchin chars to internally used wakuchin chars. This is the inverse of chars_to_wakuchin.