Function wakuchin::convert::chars_to_wakuchin

pub fn chars_to_wakuchin(chars: &str) -> String

Convert from internally used wakuchin chars to actual wakuchin chars. This is useful when you want to display the wakuchin chars to the user.

Arguments

chars - internal wakuchin chars to convert

Returns

String - actual wakuchin chars

Examples

use wakuchin::convert::chars_to_wakuchin;

assert_eq!(chars_to_wakuchin("WKCN"), "わくちん");
assert_eq!(chars_to_wakuchin("WKNCWKNC"), "わくんちわくんち");