Function wakuchin::worker::run_par

pub async fn run_par<F>(
    tries: usize, 
    times: usize, 
    regex: Regex, 
    progress_handler: F, 
    interval: Duration, 
    workers: usize
) -> Result<WakuchinResult, NormalError> where
    F: Fn(&[Progress], &[HitCounter], Duration, usize, bool) + Copy + Send + Sync + 'static, 

Research wakuchin with parallelism.

tries - number of tries

times - wakuchin times n

regex - regular expression to detect hit

progress_handler - handler function to handle progress

progress - progress information of all workers

hit_counter - hit counter, see HitCounter

interval - interval of progress refresh you were specified, usable for calculation of speed

current_diff - current - previous of progress, usable for calculation of speed

all_done - if all workers are done, usable for finalization

workers - number of workers you want to use, default to number of logical cores

Returns

WakuchinResult - the result of the research

Examples

use std::time::Duration;

use regex::Regex;

use wakuchin::worker::run_par;

let result = run_par(10, 1, Regex::new(r"WKCN").unwrap(), |_, counters, _, _, _| {
  println!("total hits: {}", counters.iter().map(|c| c.hits).sum::<usize>());
}, Duration::from_secs(1), 4);