Crate wakuchin

Core functions of wakuchin tools

Modules

builder

convert

Wakuchin conversion functions

error

handlers

progress

result

Functions to manipulate the result of a research

symbol

Wakuchin symbol definitions

worker

Wakuchin researcher main functions

Functions

check

Check wakuchin string with specified regular expression. This function is a wrapper of Regex::is_match.

gen

Generate a randomized wakuchin string.

gen_vec

Generate a vector of randomized wakuchin string. This function is useful when you want to generate multiple wakuchin strings.

validate

Check if a string is a internally used wakuchin string.

validate_external

Check whether the string is a valid wakuchin characters.