Function wakuchin::worker::run_par

pub fn run_par(
    tries: usize,
    times: usize,
    regex: &Regex,
    progress_handler: Box<dyn ProgressHandler>,
    progress_interval: Duration,
    workers: usize
) -> Result<WakuchinResult, WakuchinError>

Research wakuchin with parallelism.

Arguments

tries - number of tries If you passed zero, this function do nothing and return immediately with an empty WakuchinResult.

times - wakuchin times n, cannot be zero

regex - compiled regular expression to detect hit

progress_handler - handler function to handle progress

progress_interval - progress refresh interval

workers - number of workers you want to use, default to number of logical cores

Returns

Result<WakuchinResult, WakuchinError> - the result of the research (see WakuchinResult)

Errors

WakuchinError::TimesIsZero - Returns if you passed a zero to times

use std::sync::Mutex;
use std::time::Duration;

use regex::Regex;

use wakuchin::handlers::ProgressHandler;
use wakuchin::handlers::empty::EmptyProgressHandler;
use wakuchin::worker::run_par;

let handler: Box<dyn ProgressHandler> = Box::new(EmptyProgressHandler::new());
let result = run_par(10, 0, &Regex::new(r"WKCN")?, handler, Duration::from_secs(1), 0);

assert!(result.is_err());
assert_eq!(result.err().unwrap().to_string(), "times cannot be zero");

WakuchinError::WorkerError - Returns when any worker raised an error

Examples

use std::io::stdout;
use std::sync::{Arc, Mutex};
use std::time::Duration;

use regex::Regex;

use wakuchin::handlers::ProgressHandler;
use wakuchin::handlers::msgpack::MsgpackProgressHandler;
use wakuchin::result::{out, ResultOutputFormat};
use wakuchin::worker::run_par;

let tries = 10;
let handler: Box<dyn ProgressHandler>
  = Box::new(MsgpackProgressHandler::new(tries, Arc::new(Mutex::new(stdout()))));
let result = run_par(tries, 1, &Regex::new(r"WKCN")?, handler, Duration::from_secs(1), 4)?;

println!("{}", result.out(ResultOutputFormat::Text)?);