Function wakuchin::worker::run_par

pub async fn run_par<F>(
    tries: usize, 
    times: usize, 
    regex: Regex, 
    progress_handler: F, 
    progress_interval: Duration, 
    workers: usize
) -> Result<WakuchinResult, Box<dyn Error>> where
    F: Fn(&[Progress], &[HitCounter], Duration, usize, bool),
    F: Sync + Send + 'static, 

Research wakuchin with parallelism.

Arguments

tries - number of tries
If you passed zero, this function do nothing and return immediately with an empty WakuchinResult.

times - wakuchin times n, cannot be zero

regex - compiled regular expression to detect hit

progress_handler - handler function to handle progress

progress - referenced slice of Progress

counters - referenced slice of HitCounter

interval - interval of progress refresh you were specified, usable for calculation of speed

current_diff - current - previous of progress, usable for calculation of speed

all_done - true if all workers are done, usable for finalization

progress_interval - progress refresh interval

workers - number of workers you want to use, default to number of logical cores

Returns

Result<WakuchinResult, Box<dyn Error>> - the result of the research (see WakuchinResult)

Errors

Error::TimesIsZero - Returns if you passed a zero to times

use std::time::Duration;

use regex::Regex;

use wakuchin::worker::run_par;

let result = run_par(10, 0, Regex::new(r"WKCN")?, |_, _, _, _, _| {}, Duration::from_secs(1), 0).await;

assert!(result.is_err());
assert_eq!(result.err().unwrap().to_string(), "times cannot be zero");

JoinError - Returns when any worker raised an error

Examples

use std::time::Duration;

use regex::Regex;

use wakuchin::result::{out, ResultOutputFormat};
use wakuchin::worker::run_par;

let result = run_par(10, 1, Regex::new(r"WKCN")?, |_, counters, _, _, _| {
  println!("total hits: {}", counters.iter().map(|c| c.hits).sum::<usize>());
}, Duration::from_secs(1), 4).await?;

println!("{}", result.out(ResultOutputFormat::Text)?);